Move TerrainManifest code
